legacy_python/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
passlib==1.7.4
bcrypt==3.2.0     
email-validator==2.1.0.post1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6